refactoring rules implementation to the visibility backend.
